ansible/roles/cls/requirements-test.txt
# Testing requirements for CLS role
molecule>=3.0.0
molecule-docker>=1.0.0
testinfra>=5.0.0
pytest>=6.0.0
ansible-lint>=5.0.0
yamllint>=1.26.0
docker>=5.0.0

